+++ vector/src/Data/Vector/Mutable.hs
@@ -57,6 +57,7 @@ module Data.Vector.Mutable (
   ifoldr, ifoldr', ifoldrM, ifoldrM',
 
   -- * Modifying vectors
+  mapInPlace, imapInPlace, mapInPlaceM, imapInPlaceM,
   nextPermutation, nextPermutationBy,
   prevPermutation, prevPermutationBy,
 
@@ -570,6 +571,34 @@ unsafeMove = G.unsafeMove
 
 -- Modifying vectors
 -- -----------------
+
+-- | Modify vector in place by applying function to each element.
+--
+-- @since NEXT_VERSION
+mapInPlace :: (PrimMonad m) => (a -> a) -> MVector (PrimState m) a -> m ()
+{-# INLINE mapInPlace #-}
+mapInPlace = G.mapInPlace
+
+-- | Modify vector in place by applying function to each element and its index.
+--
+-- @since NEXT_VERSION
+imapInPlace :: (PrimMonad m) => (Int -> a -> a) -> MVector (PrimState m) a -> m ()
+{-# INLINE imapInPlace #-}
+imapInPlace = G.imapInPlace
+
+-- | Modify vector in place by applying monadic function to each element in order.
+--
+-- @since NEXT_VERSION
+mapInPlaceM :: (PrimMonad m) => (a -> m a) -> MVector (PrimState m) a -> m ()
+{-# INLINE mapInPlaceM #-}
+mapInPlaceM = G.mapInPlaceM
+
+-- | Modify vector in place by applying monadic function to each element and its index in order.
+--
+-- @since NEXT_VERSION
+imapInPlaceM :: (PrimMonad m) => (Int -> a -> m a) -> MVector (PrimState m) a -> m ()
+{-# INLINE imapInPlaceM #-}
+imapInPlaceM = G.imapInPlaceM
 
 -- | Compute the (lexicographically) next permutation of the given vector in-place.
 -- Returns False when the input is the last item in the enumeration, i.e., if it is in